chore(logs): route facet drop-rule annotation through query runner
Keep the presentation->facade import contract intact: the facet_values
view now reaches the drop-rule annotation via the allow-listed
log_facet_values_query_runner instead of importing drop_rule_matching
directly. drop_rule_matching stays the pure, shared matcher.

Generated-By: PostHog Code
Task-Id: a2f2007c-c797-46f1-8bcf-c9f536bcb393
